--- src/main/java/org/foundationsjala/coding/Library.java
@@ -1,16 +1,0 @@
-package org.foundationsjala.coding;
-
-/**
- * This Java source file was auto generated by running 'gradle buildInit --type java-library'.
- */
-public class Library {
-
-    /**
-     * Some library method.
-     *
-     * @return <code>true</code>
-     */
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
