--- src/main/java/org/foundationsjala/coding/movies/Rental.java
@@ -1,21 +1,0 @@
-package org.foundationsjala.coding.movies;
-
-public class Rental {
-
-  private Movie _movie;
-  private int _daysRented;
-
-  public Rental(Movie movie, int daysRented) {
-    _movie = movie;
-    _daysRented = daysRented;
-  }
-
-  public int getDaysRented() {
-    return _daysRented;
-  }
-
-  public Movie getMovie() {
-    return _movie;
-  }
-
-}
