--- src/main/java/org/foundationsjala/coding/movies/Type.java
@@ -1,14 +1,0 @@
-package org.foundationsjala.coding.movies;
-
-public class Type {
-  /*public static final int CHILDRENS = 2;
-  public static final int REGULAR = 0;
-  public static final int NEW_RELEASE = 1;
-  * */
-  private Type status;
-  public enum TypeStatus {
-    CHILDRENS,
-    REGULAR,
-    NEW_RELEASE;
-  }
-}
