+++ src/main/java/org/fundacionjala/coding/Library.java
@@ -1,0 +1,16 @@
+package org.fundacionjala.coding;
+
+/**
+ * This Java source file was auto generated by running 'gradle buildInit --type java-library'.
+ */
+public class Library {
+
+    /**
+     * Some library method.
+     *
+     * @return <code>true</code>
+     */
+    public boolean someLibraryMethod() {
+        return true;
+    }
+}
