--- src/test/java/org/foundationsjala/coding/LibraryTest.java
@@ -1,20 +1,0 @@
-package org.foundationsjala.coding;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-/**
- * Test for {@link Library}.
- */
-public class LibraryTest {
-
-    /**
-     * Verify some library method.
-     */
-    @Test
-    public void testSomeLibraryMethod() {
-        Library classUnderTest = new Library();
-        assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
-    }
-}
