+++ src/test/java/org/fundacionjala/coding/LibraryTest.java
@@ -1,0 +1,20 @@
+package org.fundacionjala.coding;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertTrue;
+
+/**
+ * Test for {@link Library}.
+ */
+public class LibraryTest {
+
+    /**
+     * Verify some library method.
+     */
+    @Test
+    public void testSomeLibraryMethod() {
+        Library classUnderTest = new Library();
+        assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
+    }
+}
